Split test_packages_and_services.py in package_install scenario into two files
